requirements.txt
torch
torchvision
numpy
Pillow
insightface
scikit-learn
requests
kaggle